ampcontext example now working. Created context util func.
Modified the A4A and 3P path to pass sentinel directly when
setting up listeners, instead of attaching it as data attrib on
the iframe. Got example working.

Created a utility function to get context metadata set.

Had to modify the creation of pageViewId slightly, so that it can
be called directly rather than through a class constructor, that
way it can be made and passed on to an iframe's name attrib
with the other context metadata.

Pass sentinel in to postMessageToWindows

Updated test

pageViewId must be created before iframe attached

Fixed lint issues